modified plotting by aircraft type
